+++ app/field/ArrayOfPixels.cpp
@@ -2,11 +2,11 @@
 
 int ArrayOfPixels::width() const
 {
-	return size();
+	return static_cast<int>(size());
 }
 
 int ArrayOfPixels::height() const
 {
 	if (size() == 0) return 0;
-	return (at(0).size());
+	return static_cast<int>(at(0).size());
 }
